Improve navigation drawer
This reverts to the intended "only use 2/3 of the width on narrow
screens" behaviour. (I appear to have introduced a bug when converting
to LocalWindowInfo, although I could have sworn I backed that change out
when I noticed this didn't work.) It also switches to a newer
ModalDrawerSheet version which means we get automatic back handling, so
pressing back with the navigation drawer open now closes the drawer
instead of going back from the home screen and thereby exiting the app.